contrib/SegmentAnything/requirements.txt
pyyaml >= 5.1
visualdl >= 2.2.0
opencv-python <= 4.6.0
tqdm
filelock
scipy
prettytable
scikit-learn
ftfy
regex